+++ Carmunication/Controllers/LoginViewController.swift
@@ -16,66 +16,17 @@ final class LoginViewController: UIViewController {
     // 애플 로그인 파이어베이스 인증 시 재전송 공격을 방지하기 위해 요청에 포함시키는 임의의 문자열 값
     private var currentNonce: String?
 
+    private let loginView = LoginView()
+
     override func viewDidLoad() {
         super.viewDidLoad()
         self.view.backgroundColor = .systemBackground
-        setLoginViewLayout()
-    }
 
-    // MARK: - 로그인 뷰 레이아웃 세팅
-    private func setLoginViewLayout() {
-        // MARK: - 로고
-        let logo = UIImageView(image: UIImage(systemName: "car"))
-        self.view.addSubview(logo)
-
-        // MARK: - 로고 위 텍스트
-        let logoUpperLabel = UILabel()
-        logoUpperLabel.text = "여정을 연결하다."
-        logoUpperLabel.font = UIFont.systemFont(ofSize: 22)
-        self.view.addSubview(logoUpperLabel)
-
-        // MARK: - 로고 아래 텍스트
-        let logoLowerLabel = UILabel()
-        logoLowerLabel.text = "Carmu"
-        logoLowerLabel.font = UIFont.systemFont(ofSize: 36)
-        self.view.addSubview(logoLowerLabel)
-
-        // MARK: - 애플 로그인 버튼
-        let appleSignInButton = ASAuthorizationAppleIDButton(type: .signIn, style: .black)
-        appleSignInButton.addTarget(self, action: #selector(startSignInWithAppleFlow), for: .touchUpInside)
-        appleSignInButton.cornerRadius = 20
-        self.view.addSubview(appleSignInButton)
-
-        // MARK: - 하단 회사명
-        let corpLabel = UILabel()
-        corpLabel.text = "@Damn Good Company"
-        corpLabel.font = UIFont.systemFont(ofSize: 12)
-        self.view.addSubview(corpLabel)
-
-        // MARK: - 오토 레이아웃 세팅
-        logo.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.bottom.equalTo(appleSignInButton.snp.top).offset(-202)
-            make.width.equalTo(282)
-            make.height.equalTo(141)
+        view.addSubview(loginView)
+        loginView.snp.makeConstraints { make in
+            make.edges.equalToSuperview()
         }
-        logoUpperLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.bottom.equalTo(logo.snp.top).offset(-16)
-        }
-        logoLowerLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(logo.snp.bottom).offset(20)
-        }
-        appleSignInButton.snp.makeConstraints { make in
-            make.bottom.equalTo(corpLabel.snp.top).offset(-64)
-            make.leading.trailing.equalToSuperview().inset(60)
-            make.height.equalTo(50)
-        }
-        corpLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.bottom.equalTo(view.safeAreaLayoutGuide).inset(16)
-        }
+        loginView.appleSignInButton.addTarget(self, action: #selector(startSignInWithAppleFlow), for: .touchUpInside)
     }
 }
 
@@ -136,7 +87,7 @@ extension LoginViewController: ASAuthorizationControllerDelegate {
     func authorizationController(controller: ASAuthorizationController, didCompleteWithError error: Error) {
         print("애플 로그인 실패: \(error.localizedDescription)")
     }
-    
+
     private func saveUserInKeychain(_ userIdentifier: String) {
         do {
             try KeychainItem(service: "com.dgc.carmunication", account: "userIdentifier").saveItem(userIdentifier)
